Create king-Qos
Contained in this post is a free copy of my  QoS tree (compatible with v6.0) for anyone to do what they want with, I only ask that if you republish this you include a link to this post.
https://github.com/stableagent/Firewall/
It is intended to work on a per-interface basis; with you specifying the WAN interface and the speed limit it is to have. You can then use simple-queues for your internal users for a full double-qos solution.


Please note:
1. This does not utilise connection marking so is CPU inefficient
2. This does not use any Layer7 filtering
3. This is highly focused on Residential service offerings (prioritising games etc)

Please feel free to submit your comments, questions or additions in the comments below. I plan on releasing a few new posts in the next month that cover P2P filtering as well as a newer QoS tree dealing with all the traffic below as well as P2P. There are a number of additional things missing from this early attempt that I’d welcome you to try and find, but if you’re new to QoS I feel it’s an excellent starting point 🙂

For your usage you will need to modify line 33 (and/or replicate) to add your own server ranges/admin ranges and things you want to give high priority, while the “site-specific” address list can be used to add servers or IP blocks requested for higher priority by end users (Teamspeak, Mumble and Ventrilo servers are common for this).

You may also wish to modify queue priorities to better match your network requirements (I’m sure someone out there wants to prioritise VPNs over games!)


This script uses a nested queue tree breaking all traffic into the following categories.

Download

Download Interactive
Download Non-Interactive
Upload

Upload Interactive
Upload Non-Interactive
This breakdown is similar to that found in some netequalizer, packeteer, and other QoS appliances and was first bought to my attention by Butch Evans

This allows the allocation of a top level bandwidth (say 100M) then a breakdown of this bandwidth to an interactive (90M web traffic, skype, things that users are sitting watching/waiting for) traffic queue and a non-interactive (10M torrents, background downloads, windows updates) traffic queue with a lower priority. Using PCQ queues it also permits bandwidth to be equally shared among users and allows the lower priority queues to burst up and make use of the other bandwidth as available.

For the download traffic queues I would recommend setting your values to between 90%-98% of your inbound bandwidth to ensure you are delaying/queuing traffic before it is being queued or policed upstream.